+++ src/main/java/services/Service.java
@@ -1,4 +1,4 @@
-package Services;
+package services;
 
 public interface Service {
     public void printElements();
